Temporary fix for deletebundle's default behavior
When euca-delete-bundle is supplied with neither a manifest path nor a
key prefix the tool deletes all bundles in the entire bucket.  This
behavior is deprecated and not user-friendly, but we don't want to break
scripts in a z-stream release, so we compromise:  before
euca-delete-bundle deletes everything it will pause for ten seconds to
give the user a chance to ^C out of it.

The upcoming major release will either require one to explicitly opt
into this behavior with a command line switch or remove it altogether.