Drop default-region from our included config file
Setting aws:us-east-1 as euca2ools' default region sounds reasonable at
first, but when people use environment variables (a.k.a. eucarc) as
their sources of configuration thish causes commands to connect to AWS
when Eucalyptus-equivalent URLs for the services they access are not
defined.  The result is a semi-misleading "The AWS Access Key Id you
provided does not exist in our records" message, so here we just drop it
to prevent this from confusing people who use the default configuration.

Fixes TOOLS-393