Remove eustore client tools
The current version of eustore requires too much cohesion and is so
inefficient when operating that it really needs a complete rewrite.
That said, as eucalyptus is moving toward HVM-based images, perhaps
said rewrite will not be necessary.  Either way, it will not be based
on this code.  See https://eucalyptus.atlassian.net/browse/TOOLS-423
for more info.

Fixes TOOLS-423