Make GetConsoleOutput fall back to utf-8 encoding
When sys.stdout points to a file its encoding is None, so this change
fixes encoding issues for this specific case (TOOLS-407).  This is only
one instance of a general problem with output encoding, though, which
must be fixed at the framework level, which is not feasible in a patch
release.  See TOOLS-385 for the more general issue.

Fixes TOOLS-407