Add euare-instanceprofilecreate --create-role
A common workflow for instance profile creation is as follows:
  1. euare-rolecreate -r foo
  2. euare-instanceprofilecreate -s foo -r foo

This process winds up with typos far more frequently than one might
initially expect.  Rather than adding an option to
euare-instanceprofilecreate that causes it to check whether the role
exists, we instead add an option that automates the whole workflow by
making that command create a role with the same name and add it to the
newly-created instance profile.

Fixes TOOLS-463