Fix proxy handling for euca-* commands
Requestbuilder commit a6c4d64 fixes HTTP_PROXY and HTTPS_PROXY handling
for all but euca-* commands, which rely on their own copy of the method
that was fixed.  This commit applies the fix to that copy.

To fix this problem for non-euca-* commands, use requestbuilder 0.1.x >=
0.1.1 or requestbuilder 0.2.x >= 0.2.2.

Fixes TOOLS-447 only for euca-* commands