Deal with bad parsing of boolean filter values
The bool() function evaluates every non-empty string as True, breaking
args that use that as their types.  Strictly speaking, this commit just
just sidesteps the problem for the one filter that has this problem
(block-device-mapping.delete-on-termination) by changing its type to
str.  As the server handles all filter values with plain string
comparisons this is reasonable, and, in fact, should be expected for all
filters.

In case this comes up for regular args for some reason, however, this
commit also adds an arg type called flexible_bool that properly
interprets a variety of common strings for boolean values.

Fixes TOOLS-502