Make BundleInstance attempt to draw creds from config
This commit makes euca-bundle-instance attempt to get access keys and
secret keys from configuration files if they are not supplied at the
command line.  It does not attempt to use the same credentials as the
auth handler that is being used for EC2 because we may be using STS
credentials for the latter.  S3 does not appear to support STS
credentials in POST upload policies.  Moreover, we cannot determine when
STS credentials will expire; if they were to expire while a bundle task
is running the task will fail inexplicably.

Fixes TOOLS-478