Add euare-usercreate --verify for pre-flight checking
Since euare-usercreate can be called with a group name that does not
exist, one can end up in a state of partial failure, where re-running
the command fails because the new user already exists.  The new --verify
option makes the command check for the group's existence before running.
This is not the default behavior because this command must not require
permissions beyond iam:CreateUser in the default case.

Fixes TOOLS-534