Warn and fake a region name when one is missing
HmacV4Auth requires a region name in order to function, but the eucarc
files that many people use today lack them.  For the time being we
therefore add a fake region name when one is not otherwise present and
print a warning with instructions on how to supply one.  This will not
work against AWS, but euca2ools ship with config file entries for that,
hopefully limiting the number of people who will encounter that case.

This hack should be removed some time after euca2ools 3.3.

Completes TOOLS-360