Add eu-central-1, use HmacV4Auth against it and cn-north-1
Eucalyptus doesn't yet support HmacV4Auth for its S3 endpoint, so we
instead trigger that signature mechanism only when a new
"s3-force-sigv4" region option is set and then flip that on for the AWS
regions that require it.  This will be removed in the future once
eucalyptus has supported that signature method for a period of time.

Fixes TOOLS-469
Fixes TOOLS-524