Add a --check-cert option to bundle commands
This commit adds a --check-cert option to several bundle-related
commands that makes them run through each signing certificate for each
user in the account in an attempt to find one with the same fingerprint
as the one supplied to the bundling process.  If a match is not found or
a match is found, but the certificate is inactive, the commands will
abort and show an appropriate error message.

This will not work for users without permission to access the IAM
service.

Fixes TOOLS-550