Add key=value attr syntax to eulb-create-lb-policy
The syntax for supplying attributes to eulb-create-lb-policy is
ridiculous and makes creating a policy with a large number of attributes
for things like crypto settings unreasonably annoying.  This commit adds
a new option, -A, that allows one to use comma-separated lists of
key-value pairs that are themselves separated by "=" instead.

Fixes TOOLS-551