Add --write-config to CreateAccessKey/User/Account
A common activity for a cloud admin is to create a file containing
an access key and URLs for all of the cloud's services so he or she can
send it to an end user.  The new -w switch for euare commands that can
create access keys causes it to write those access keys as well as
service URLs in the form of euca2ools.ini(5) files to make that process
as seamless as possible.  This can act as  a client-side replacement for
euca_conf --get-credentials.

    euare-useraddkey -wd cloud.example.com

    euare-usercreate -u someuser -wd cloud.example.com

    euare-accountcreate -wd cloud.example.com

Implements EUCA-10590