Log a warning when only one of -U and --ec2-url is used
Running euca-install-image with -U/--url only updates the S3 URL it
uses, not the EC2 URL.  If an EC2 URL is available from configuration or
the environment this could result in it talking to two different regions
and the assortment of confusing issues that can result, so when only one
of -U/--url and --ec2-url is used we now log a warning to help identify
this case.

See EUCA-10984 for an example of the case this addresses.