Make all bundle certs/keys optional
This commit makes user certificates and private keys optional by simply
omitting or otherwise filling in dummy values for the parts of bundle
manifests that they normally produce.  Note that unsigned bundles
require eucalyptus 4.2 and certificate-less bundles cannot be unbundled
by the user.  The cloud certificate continues to be mandatory, but is
fetched from the empyrean service if possible.

This also removes the --check-cert option, as it is now obsolete.

Fixes TOOLS-572
Reverts TOOLS-550
Requires EUCA-10784
Requires EUCA-10838