Add euca-generate-environment-config
This command reads the environment variables and configuration file
variables for each service and each bit of auth information that
euca2ools commands read and then outputs them as shellcode that sets
them all as environment variables.  The uses for this are twofold:

1.  Generating eucarc-like files for legacy scripts
2.  Making it possible for eucalyptus code that does not use
    requestbuilder to leverage euca2ools's configuration mechanisms

In the latter case, the --simple option can simplify the output one
needs to parse.