Add -l/--set-default-user to CreateAccessKey
To make things a little easier for first-time users this commit adds a
-l option to euare-useraddkey and other commands that generate
euca2ools.ini files that causes them to set the user it generates as the
default for the region.  This is particularly useful for administrators
who are generating people's accounts and users for the first time and
distributing their credentials.

Example:  euare-accountcreate -wld mycloud.example.com > creds.ini